+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changes
 
-## v4.0.0 (XXX XX, 2022)
+## v4.0.0 (February XX, 2022)
 ### Features 🔊
 - Render environments to WanDB
 - Grouping of runs in WanDB
@@ -11,7 +11,6 @@
 - LogCosh instead of Huber loss
 - Critic network with Add layer applied on state & action branches
 - Custom uniform initializer
-- [LayerNormalization + TanH] block for Input layer
 - XLA (Accelerated Linear Algebra) compiler
 - Optimized Replay Buffer (https://github.com/deepmind/reverb/issues/90)
 - split into **Agent**, **Learner**, **Tester** and **Server**